Fix build

+++ src/rpc/methods/eth.rs
@@ -1078,7 +1078,7 @@ async fn new_eth_tx_receipt<DB: Blockstore>(
         block_hash: tx.block_hash.clone(),
         block_number: tx.block_number.clone(),
         r#type: tx.r#type.clone(),
-        status: (message_lookup.receipt.exit_code().is_success() as u8).into(),
+        status: (message_lookup.receipt.exit_code().is_success() as u64).into(),
         gas_used: message_lookup.receipt.gas_used().into(),
         ..EthTxReceipt::new()
     };
